luce_vm/luce_django/luce/requirements.txt
Brownie==0.5.1
Django==4.1.12
djangorestframework==3.14.0
hexbytes==0.2.3
hexbytes==0.2.0
matplotlib==3.5.1
py_solc_x==1.1.1
Requests==2.31.0
urllib3==1.26.12
urllib3==1.26.14
urllib3==1.25.3
